requirements.txt
data-repository
wheel
setuptools
polars
pyarrow
pandas
kagglehub
datasets
polars